Added fetch france to index.d.ts
