+++ lib/dgtresor_gouv_fr__national_freeze_registry.d.ts
@@ -1,0 +1,3 @@
+export declare function dgtresor_gouv_fr__national_freeze_registry(): Promise<{
+    [key: string]: any;
+}[]>;
