--- lib/fetch_uk.d.ts
@@ -1,3 +1,0 @@
-export declare function fetchUK(): Promise<{
-    [key: string]: any;
-}[]>;
