--- lib/fetch_usa___interpol.d.ts
@@ -1,5 +1,0 @@
-/**
- * Accesses Interpol Notices API and retrieves Red Notice data
- * @returns {any[]} Array of JS Objects with red notice data
- */
-export declare function fetchIP(): Promise<any[]>;
