--- lib/fetch_wb.d.ts
@@ -1,3 +1,0 @@
-export declare function fetchWB(): Promise<{
-    [key: string]: any;
-}[]>;
