--- lib/fiu_gov_ua__black_list_full.d.ts
@@ -1,3 +1,0 @@
-export declare function fiu_gov_ua__black_list_full(): Promise<{
-    [key: string]: any;
-}[]>;
