--- lib/gc_ca__casl.d.ts
@@ -1,3 +1,0 @@
-export declare function gc_ca__casl(): Promise<{
-    [key: string]: any;
-}[]>;
