--- lib/government_nl__dnslt.d.ts
@@ -1,3 +1,0 @@
-export declare function government_nl__dnslt(): Promise<{
-    [key: string]: any;
-}[]>;
