+++ lib/index.js
@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.fiu_gov_ua__terrorist_activity = exports.interpol_int__red_notices = exports.justice_gc_ca__corrupt_foreign_nationals = exports.gc_ca__consol_autonomous_sanctions = exports.occ_gov__enforcement_actions = exports.bis_doc_gov__denied_persons = exports.api_trade_gov__csl_consolidated = exports.treasury_gov__nonsdnl = exports.pmddtc_state_gov__aeca_dsl = exports.un_org__consolidated_entities = exports.un_org__consolidated_individuals = exports.worldbank_org__debarred_firms_and_individuals = exports.government_nl__dutch_national_sanctions_list = exports.ec_europa_eu__sanctions_list = exports.gov_uk__financial_sanctions_list = exports.dfat_gov_au__consolidated_list = void 0;
+exports.fiu_gov_ua__terrorist_activity = exports.interpol_int__red_notices = exports.dgtresor_gouv_fr__national_freeze_registry = exports.justice_gc_ca__corrupt_foreign_nationals = exports.gc_ca__consol_autonomous_sanctions = exports.occ_gov__enforcement_actions = exports.bis_doc_gov__denied_persons = exports.api_trade_gov__csl_consolidated = exports.treasury_gov__nonsdnl = exports.pmddtc_state_gov__aeca_dsl = exports.un_org__consolidated_entities = exports.un_org__consolidated_individuals = exports.worldbank_org__debarred_firms_and_individuals = exports.government_nl__dutch_national_sanctions_list = exports.ec_europa_eu__sanctions_list = exports.gov_uk__financial_sanctions_list = exports.dfat_gov_au__consolidated_list = void 0;
 var dfat_gov_au__consolidated_list_1 = require("./dfat_gov_au__consolidated_list");
 Object.defineProperty(exports, "dfat_gov_au__consolidated_list", { enumerable: true, get: function () { return dfat_gov_au__consolidated_list_1.dfat_gov_au__consolidated_list; } });
 var gov_uk__financial_sanctions_list_1 = require("./gov_uk__financial_sanctions_list");
@@ -33,11 +33,13 @@ var occ_gov__enforcement_actions_1 = require("./occ_gov__enforcement_actions");
 Object.defineProperty(exports, "occ_gov__enforcement_actions", { enumerable: true, get: function () { return occ_gov__enforcement_actions_1.occ_gov__enforcement_actions; } });
 // ecfr_gov__militaryenduser
 // ecfr_gov__unverified
-// dgtresor_gouv_fr__national_freeze_registry
 var gc_ca__consol_autonomous_sanctions_1 = require("./gc_ca__consol_autonomous_sanctions");
 Object.defineProperty(exports, "gc_ca__consol_autonomous_sanctions", { enumerable: true, get: function () { return gc_ca__consol_autonomous_sanctions_1.gc_ca__consol_autonomous_sanctions; } });
 var justice_gc_ca__corrupt_foreign_nationals_1 = require("./justice_gc_ca__corrupt_foreign_nationals");
 Object.defineProperty(exports, "justice_gc_ca__corrupt_foreign_nationals", { enumerable: true, get: function () { return justice_gc_ca__corrupt_foreign_nationals_1.justice_gc_ca__corrupt_foreign_nationals; } });
+//export { gc_ca__casl } from "./gc_ca__casl";
+var dgtresor_gouv_fr__national_freeze_registry_1 = require("./dgtresor_gouv_fr__national_freeze_registry");
+Object.defineProperty(exports, "dgtresor_gouv_fr__national_freeze_registry", { enumerable: true, get: function () { return dgtresor_gouv_fr__national_freeze_registry_1.dgtresor_gouv_fr__national_freeze_registry; } });
 //justice_gc_ca__cfn_tunisia
 //justice_gc_ca__cfn_ukraine
 //publicsafety_gc_ca__counter_terrorism_entity
