--- lib/occ_gov_enforcement_actions.d.ts
@@ -1,3 +1,0 @@
-export declare function occ_gov_enforcement_actions(): Promise<{
-    [key: string]: any;
-}[]>;
