+++ lib/runner.js
@@ -14,6 +14,8 @@ const occ_gov__enforcement_actions_1 = require("./occ_gov__enforcement_actions")
 const gc_ca__consol_autonomous_sanctions_1 = require("./gc_ca__consol_autonomous_sanctions");
 const treasury_gov__nonsdnl_1 = require("./treasury_gov__nonsdnl");
 const api_trade_gov__csl_consolidated_1 = require("./api_trade_gov__csl_consolidated");
+//import { fiu_gov_ua__black_list_full } from "./fiu_gov_ua__black_list_full";
+const dgtresor_gouv_fr__national_freeze_registry_1 = require("./dgtresor_gouv_fr__national_freeze_registry");
 const fiu_gov_ua__terrorist_activity_1 = require("./fiu_gov_ua__terrorist_activity");
 if (argv.au !== undefined) {
     console.log("fetch au list...");
@@ -66,5 +68,9 @@ if (argv.usacslapi !== undefined) {
 if (argv.fiuua !== undefined) {
     console.log("fetch ukraine list...");
     (0, fiu_gov_ua__terrorist_activity_1.fiu_gov_ua__terrorist_activity)();
+}
+if (argv.frnfr !== undefined) {
+    console.log("fetch france nfr list...");
+    (0, dgtresor_gouv_fr__national_freeze_registry_1.dgtresor_gouv_fr__national_freeze_registry)();
 }
 //# sourceMappingURL=runner.js.map
